+++ GoogleMapsTestApp/AppDelegate.swift
@@ -18,7 +18,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
         // Override point for customization after application launch.
  
-        GMSServices.provideAPIKey(APIKey.key)
+        GMSServices.provideAPIKey(GoogleMapsUtilsAPIKey.value)
         Launcher.launch(with: window)
         
         return true
