+++ GoogleMapsTestApp/ClusterConfigurator/ClusterManagerConfigurator.swift
