+++ GoogleMapsTestApp/ClusterConfigurator/GMUComponentFactory.swift
