+++ GoogleMapsTestApp/ClusterConfigurator/GMUComponentFactoryProtocol.swift
