+++ GoogleMapsTestApp/Constants/StartPoint.swift
