+++ GoogleMapsTestApp/Protocols/ClusterConfiguratorDataSourceProtocol.swift
