+++ GoogleMapsTestApp/Protocols/ClusterConfiguratorProtocol.swift
