+++ GoogleMapsTestApp/Protocols/MapPointCategoryContainerType.swift
@@ -1,0 +1,13 @@
+//
+//  MapPointCategoryContainerType.swift
+//  GoogleMapsTestApp
+//
+//  Created by Andrey on 19.11.2019.
+//  Copyright © 2019 Andrey. All rights reserved.
+//
+
+import Foundation
+
+protocol MapPointCategoryContainerType {
+    var category: MarkerCategory { get }
+}
