+++ GoogleMapsTestApp/Reusables/POIItem/POIItem.swift
