+++ GoogleMapsTestApp/Scenes/Buttons/View/ButtonViewControllerProtocol.swift
@@ -1,0 +1,13 @@
+//
+//  ButtonViewControllerProtocol.swift
+//  GoogleMapsTestApp
+//
+//  Created by Andrey on 19.11.2019.
+//  Copyright © 2019 Andrey. All rights reserved.
+//
+
+import Foundation
+
+protocol ButtonViewControllerProtocol: class {
+    var viewModel: ButtonSceneViewModelProtocol! { get set }
+}
