+++ GoogleMapsTestApp/Scenes/GoogleMaps/ViewModel/GoogleMapsViewModelOutput.swift
