+++ GoogleMapsTestApp/UI elements/ZoomButton/ZoomButton.swift
